requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
appdirs==1.4.4
asttokens==3.0.1
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.4.0
backcall==0.2.0
beautifulsoup4==4.14.3
blinker==1.9.0
bs4==0.0.2
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.1.8
cloudpickle==3.1.2
colorama==0.4.6
comm==0.2.3
contourpy==1.3.2
cryptography==44.0.3
cycler==0.12.1
debugpy==1.8.18
decorator==5.2.1
distro==1.9.0
docker-pycreds==0.4.0
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.124.2
filelock==3.20.0
filetype==1.2.0
Flask==3.1.2
flatbuffers==25.9.23
flwr==1.24.0
fonttools==4.61.0
frozenlist==1.8.0
fsspec==2025.12.0
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.41.1
google-auth-oauthlib==1.2.3
google-pasta==0.2.0
graphviz==0.21
greenlet==3.3.0
grpcio==1.76.0
grpcio-health-checking==1.76.0
h11==0.16.0
h5py==3.15.1
haversine==2.9.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.7
ImageIO==2.37.2
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==7.1.0
ipython==8.37.0
iterators==0.0.2
itsdangerous==2.2.0
jax==0.6.2
jaxlib==0.6.2
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
jupyter_client==8.7.0
jupyter_core==5.9.1
keras==3.12.0
keras-nightly==3.12.0.dev2025100703
Keras-Preprocessing==1.1.2
kiwisolver==1.4.9
lazy_loader==0.4
libclang==18.1.1
lime==0.2.0.1
llvmlite==0.46.0
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
ml_dtypes==0.5.4
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.7.0
namex==0.1.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.13.0
numba==0.63.1
numpy==2.2.6
oauthlib==3.3.1
openai==2.11.0
opencv-python==4.12.0.88
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pathspec==0.12.1
pathtools==0.1.2
pi_heif==1.1.1
pickleshare==0.7.5
pillow==12.0.0
pillow-avif-plugin==1.5.2
platformdirs==4.5.1
polars==1.36.1
polars-runtime-32==1.36.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.2
psutil==7.1.3
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pyclipper==1.4.0
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
PyDirectInput==1.0.4
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.19.2
PyMsgBox==2.0.1
pyparsing==3.2.5
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-version==0.0.2
pytweening==1.2.0
pytz==2025.2
PyWavelets==1.8.0
pywin32==311
PyYAML==6.0.3
pyzmq==27.1.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
roboflow==1.2.11
rsa==4.9.1
safetensors==0.7.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
sentry-sdk==2.47.0
setproctitle==1.3.7
shap==0.49.1
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
slicer==0.0.8
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.45
stack-data==0.6.3
starlette==0.50.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.20.0
termcolor==3.2.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tokenizers==0.22.1
tomli==2.3.0
tomli_w==1.2.0
torch==2.9.1
torchvision==0.24.1
tornado==6.5.3
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.3
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.236
ultralytics-thop==2.0.18
urllib3==2.6.2
uvicorn==0.38.0
wcwidth==0.2.14
Werkzeug==3.1.4
wget==3.2
wrapt==2.0.1
yarl==1.22.0
zipp==3.23.0
